pyproject.toml
[project]
name = "ai-journaling-agent"
version = "0.1.0"
description = "AI agent-based journaling app for cultivating mindfulness and well-being"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "line-bot-sdk>=3.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "httpx>=0.28",
    "pytest-asyncio>=0.25",
]

[project.scripts]
inbox = "ai_journaling_agent.cli.inbox_cmd:main"
push-line = "ai_journaling_agent.cli.push_cmd:main"
journal-history = "ai_journaling_agent.cli.history_cmd:main"
checkin = "ai_journaling_agent.cli.checkin_cmd:main"
journal-today = "ai_journaling_agent.cli.today_cmd:main"
mood-report = "ai_journaling_agent.cli.mood_report_cmd:main"
retrospective = "ai_journaling_agent.cli.retrospective_cmd:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_journaling_agent"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
